fixup: uart: check for busy status in flush
Also checks the `busy` status of the UART peripheral if the TX FIFO is
non-empty. Avoids clearing the FIFO prematurely.

If the peripheral is not busy, but still has buffered data in the FIFO,
clears the FIFO. Returns `nb::Error::WouldBlock` to check the FIFO
status on the next iteration of an `nb::block` call.